pyproject.toml
[tool.poetry]
name = "occ-calculator"
version = "0.1.0"
description = ""
authors = ["AngeloGagno <angelogagno@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
sqlalchemy = "^2.0.36"
psycopg2 = "^2.9.10"
oauth2client = "^4.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
